# src/lib/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": ".M.U.F.A.R. APIs",
    "version": "1.1.11"
  },
  "security": [
    {
      "apiKey": []
    }
  ],
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "query",
        "name": "apikey"
      }
    }
  },
  "paths": {
...
    "/api/anime/doujin-search": {
      "get": {
        "tags": ["Anime"],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "URL for doujin search"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "example": {
                  "status": "Success",
                  "code": 200,
                  "author": "nama_author",
                  "data": "data_doujin"
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "example": {
                  "status": "Error",
                  "code": 500,
                  "message": "Internal Server Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/anime/doujin-ch": {
      "get": {
        "tags": ["Anime"],
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "URL for doujin search"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "example": {
                  "status": "Success",
                  "code": 200,
                  "author": "nama_author",
                  "data": "data_doujin"
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "example": {
                  "status": "Error",
                  "code": 500,
                  "message": "Internal Server Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/anime/doujin-img": {
      "get": {
        "tags": ["Anime"],
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "URL doujin get image"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "example": {
                  "status": "Success",
                  "code": 200,
                  "author": "nama_author",
                  "data": "data_doujin"
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "example": {
                  "status": "Error",
                  "code": 500,
                  "message": "Internal Server Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/anime/whatanime": {
      "get": {
        "tags": ["Anime"],
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "description": "URL of the image or video frame",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "example": {}
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "error": "Invalid parameters. URL is required."
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": {
                  "error": "Internal server error."
                }
              }
            }
          }
        }
      }
    },
    "/api/anime/nhentai-search": {
      "get": {
        "tags": ["Anime"],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "q of the image or video frame",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "example": {}
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "error": "Invalid parameters. URL is required."
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": {
                  "error": "Internal server error."
                }
              }
            }
          }
        }
      }
    },
    "/api/downloader/tiktok": {
      "get": {
        "tags": ["Downloader"],
        "parameters": [
          {
            "in": "query",
            "name": "url",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "code": {
                      "type": "integer"
                    },
                    "author": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                },
                "example": {
                  "status": "Success",
                  "code": 200,
                  "author": "xyla",
                  "data": {}
                }
              }
            }
          }
        }
      }
    },
    "/api/downloader/facebook": {
      "get": {
        "tags": ["Downloader"],
        "parameters": [
          {
            "in": "query",
            "name": "url",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "code": {
                      "type": "integer"
                    },
                    "author": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                },
                "example": {
                  "status": "Success",
                  "code": 200,
                  "author": "xyla",
                  "data": {}
                }
              }
            }
          }
        }
      }
    },
    "/api/downloader/xnxx": {
      "get": {
        "tags": ["Downloader"],
        "parameters": [
          {
            "in": "query",
            "name": "url",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "code": {
                      "type": "integer"
                    },
                    "author": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                },
                "example": {
                  "status": "Success",
                  "code": 200,
                  "author": "xyla",
                  "data": {}
                }
              }
            }
          }
        }
      }
    },
    "/api/downloader/mediafire": {
      "get": {
        "tags": ["Downloader"],
        "parameters": [
          {
            "in": "query",
            "name": "url",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "code": {
                      "type": "integer"
                    },
                    "author": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                },
                "example": {
                  "status": "Success",
                  "code": 200,
                  "author": "xyla",
                  "data": {}
                }
              }
            }
          }
        }
      }
    },
...
    "/api/search/youtube": {
      "get": {
        "tags": ["Search"],
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "description": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response with downloaded YouTube content.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "code": {
                      "type": "integer"
                    },
                    "author": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                },
                "example": {
                  "status": "Success",
                  "code": 200,
                  "author": "xyla",
                  "data": {}
                }
              }
            }
          }
        }
      }
    },
    "/api/search/xnxx": {
      "get": {
        "tags": ["Search"],
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "description": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "code": {
                      "type": "integer"
                    },
                    "author": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                },
                "example": {
                  "status": "Success",
                  "code": 200,
                  "author": "xyla",
                  "data": {}
                }
              }
            }
          }
        }
      }
    },
    "/api/search/wikipedia": {
      "get": {
        "tags": ["Search"],
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "description": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "code": {
                      "type": "integer"
                    },
                    "author": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                },
                "example": {
                  "status": "Success",
                  "code": 200,
                  "author": "xyla",
                  "data": {}
                }
              }
            }
          }
        }
      }
    },
...
  }
}